requirements.txt
langchain-google-genai
langchain-community
langchain-core
flask
GitPython
python-dotenv
faiss-cpu